Change Project Name (MexRestSample) -> (MobiledgeXRestSample)
