--- unity/MexRestSample/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,0 @@
-m_EditorVersion: 2019.1.0f2
-m_EditorVersionWithRevision: 2019.1.0f2 (292b93d75a2c)
