+++ unity/MobiledgeXRestSample/.gitignore
