+++ unity/MobiledgeXRestSample/Assets/FindCloudletButtonScript.cs
