+++ unity/MobiledgeXRestSample/Assets/FullSampleButtonScript.cs.meta
