+++ unity/MobiledgeXRestSample/Assets/LocationButtonScript.cs
