+++ unity/MobiledgeXRestSample/Assets/LocationButtonScript.cs.meta
