+++ unity/MobiledgeXRestSample/Assets/LocationService.cs
