+++ unity/MobiledgeXRestSample/Assets/LocationService.cs.meta
