+++ unity/MobiledgeXRestSample/Assets/Logos.meta
