+++ unity/MobiledgeXRestSample/Assets/Logos/MexGreen.mat.meta
