+++ unity/MobiledgeXRestSample/Assets/MexLogo.unity
