+++ unity/MobiledgeXRestSample/Assets/MexLogo.unity.meta
