+++ unity/MobiledgeXRestSample/Assets/MexSample.cs
