+++ unity/MobiledgeXRestSample/Assets/PermissionScript.cs.meta
