+++ unity/MobiledgeXRestSample/Assets/PlatformIntegration.cs
