+++ unity/MobiledgeXRestSample/Assets/Plugins.meta
