+++ unity/MobiledgeXRestSample/Assets/Plugins/dummy.txt.meta
