+++ unity/MobiledgeXRestSample/Assets/RegisterClientButtonScript.cs
