+++ unity/MobiledgeXRestSample/Assets/RegisterClientButtonScript.cs.meta
