+++ unity/MobiledgeXRestSample/Assets/VerifyLocationButtonScript.cs
