+++ unity/MobiledgeXRestSample/Assets/VerifyLocationButtonScript.cs.meta
