+++ unity/MobiledgeXRestSample/Packages/manifest.json
@@ -1,15 +1,21 @@
 {
   "dependencies": {
+    "com.unity.2d.sprite": "1.0.0",
+    "com.unity.2d.tilemap": "1.0.0",
     "com.unity.ads": "2.3.1",
-    "com.unity.analytics": "3.3.2",
+    "com.unity.analytics": "3.3.5",
     "com.unity.collab-proxy": "1.2.16",
-    "com.unity.multiplayer-hlapi": "1.0.2",
-    "com.unity.package-manager-ui": "2.1.2",
+    "com.unity.ide.rider": "1.1.4",
+    "com.unity.ide.vscode": "1.1.4",
+    "com.unity.multiplayer-hlapi": "1.0.4",
     "com.unity.purchasing": "2.0.6",
-    "com.unity.textmeshpro": "2.0.0",
-    "com.unity.timeline": "1.0.0",
+    "com.unity.test-framework": "1.1.11",
+    "com.unity.textmeshpro": "2.0.1",
+    "com.unity.timeline": "1.2.6",
+    "com.unity.ugui": "1.0.0",
     "com.unity.xr.legacyinputhelpers": "2.0.2",
     "com.unity.modules.ai": "1.0.0",
+    "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
     "com.unity.modules.assetbundle": "1.0.0",
     "com.unity.modules.audio": "1.0.0",
