+++ unity/MobiledgeXRestSample/ProjectSettings/DynamicsManager.asset
