+++ unity/MobiledgeXRestSample/ProjectSettings/GraphicsSettings.asset
