+++ unity/MobiledgeXRestSample/ProjectSettings/InputManager.asset
