+++ unity/MobiledgeXRestSample/ProjectSettings/PresetManager.asset
