+++ unity/MobiledgeXRestSample/ProjectSettings/ProjectSettings.asset
