+++ unity/MobiledgeXRestSample/ProjectSettings/QualitySettings.asset
