+++ unity/MobiledgeXRestSample/ProjectSettings/TagManager.asset
