+++ unity/MobiledgeXRestSample/ProjectSettings/UnityConnectSettings.asset
