+++ unity/MobiledgeXRestSample/ProjectSettings/VFXManager.asset
