+++ unity/MobiledgeXRestSample/ProjectSettings/XRSettings.asset
@@ -1,0 +1,10 @@
+{
+    "m_SettingKeys": [
+        "VR Device Disabled",
+        "VR Device User Alert"
+    ],
+    "m_SettingValues": [
+        "False",
+        "False"
+    ]
+}
