+++ unity/MobiledgeXRestSample/iosMexRestSample/Classes/PlatformIntegration.m
